added logo link, fixed hometab
