+++ src/components/Header/Header.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { modalLogout } from '../../redux/global/globalSlice';
 import { getIsModalLogoutOpen } from '../../redux/global/globalSlice';
-import { LogoSection } from './LogoSection/LogoSection';
+import { LogoSection } from '../LogoSection/LogoSection';
 import { AccountSection } from './AccountSection/AccountSection';
 import { ModalLogout } from '../ModalLogout/ModalLogout';
